Added Relation property support and Link to pages in rich text support (#10)
* Relations in properties are now displayed

I had to create a relation_logger singleton that saves the relations already created to avoid looping over and over again on the same bidirectional relations.

* Links to pages in pages paragraphs now display as a relation

In BlockParser, the had_children property now defaults to True because Pages do not provide the has_children propery even if it has children in the form of links in the paragraph. So now, if has_children is not provided, it will try to load the children.

The RichTextParser now is able to look for text blocks that contain a href. It will assume it is a page id and will try to retrieve the page with this id. If an error is thrown, we assume it means it was an http link and we move on.
This part could be improved by match the href to a UUID format before we send the query.

RichTextParser now needs a parent_id so we can add a new relation to the relation_logger.

Some parsers were using sets for linked_blocks and children_ids, which triggered an error when trying to extend it. I assumed it was a mistake and changed them to be lists instead.

* Moved parser.py to db_to_graph_parser.py to avoid a Pylance warning on the whole parser.py file